pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestra"
version = "0.1.0"
description = "Multi-agent orchestrator"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "tavily-python>=0.5.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "agents*"]
